requirements.txt
fancycompleter==0.7
feedparser==5.2.1
Jinja2==2.9.4
MarkupSafe==0.23
ordereddict==1.1
pdbpp==0.8.3
Pygments==2.1.3
pyrepl==0.8.4
python-dateutil==2.6.0
PyYAML==3.12
six==1.10.0
wmctrl==0.3
